ENH: allow axis argument to append / move append code to generic.py (GH8295)
`append` moved from frame.py to generic.py with docs commented.

test cases for DataFrame.append moved from tests_frame.py to tests_generic.py #1

test cases for DataFrame.append moved from tests_frame.py to tests_generic.py #2

test cases for DataFrame.append moved from tests_frame.py to tests_generic.py #3

test cases for Series.append moved from tests_series.py to tests_generic.py

`append` in generic.py passes `axis` argument. And added corresponding test case
