+++ pandas/tests/test_frame.py
@@ -6606,128 +6606,6 @@ class TestDataFrame(tm.TestCase, CheckIndexing,
         mixed2 = mixed1.convert_objects()
         assert_frame_equal(mixed1, mixed2)
 
-    def test_append_series_dict(self):
-        df = DataFrame(np.random.randn(5, 4),
-                       columns=['foo', 'bar', 'baz', 'qux'])
-
-        series = df.ix[4]
-        with  assertRaisesRegexp(ValueError, 'Indexes have overlapping values'):
-            df.append(series, verify_integrity=True)
-        series.name = None
-        with assertRaisesRegexp(TypeError, 'Can only append a Series if '
-                                'ignore_index=True'):
-            df.append(series, verify_integrity=True)
-
-        result = df.append(series[::-1], ignore_index=True)
-        expected = df.append(DataFrame({0: series[::-1]}, index=df.columns).T,
-                             ignore_index=True)
-        assert_frame_equal(result, expected)
-
-        # dict
-        result = df.append(series.to_dict(), ignore_index=True)
-        assert_frame_equal(result, expected)
-
-        result = df.append(series[::-1][:3], ignore_index=True)
-        expected = df.append(DataFrame({0: series[::-1][:3]}).T,
-                             ignore_index=True)
-        assert_frame_equal(result, expected.ix[:, result.columns])
-
-        # can append when name set
-        row = df.ix[4]
-        row.name = 5
-        result = df.append(row)
-        expected = df.append(df[-1:], ignore_index=True)
-        assert_frame_equal(result, expected)
-
-    def test_append_list_of_series_dicts(self):
-        df = DataFrame(np.random.randn(5, 4),
-                       columns=['foo', 'bar', 'baz', 'qux'])
-
-        dicts = [x.to_dict() for idx, x in df.iterrows()]
-
-        result = df.append(dicts, ignore_index=True)
-        expected = df.append(df, ignore_index=True)
-        assert_frame_equal(result, expected)
-
-        # different columns
-        dicts = [{'foo': 1, 'bar': 2, 'baz': 3, 'peekaboo': 4},
-                 {'foo': 5, 'bar': 6, 'baz': 7, 'peekaboo': 8}]
-        result = df.append(dicts, ignore_index=True)
-        expected = df.append(DataFrame(dicts), ignore_index=True)
-        assert_frame_equal(result, expected)
-
-    def test_append_empty_dataframe(self):
-
-        # Empty df append empty df
-        df1 = DataFrame([])
-        df2 = DataFrame([])
-        result = df1.append(df2)
-        expected = df1.copy()
-        assert_frame_equal(result, expected)
-
-        # Non-empty df append empty df
-        df1 = DataFrame(np.random.randn(5, 2))
-        df2 = DataFrame()
-        result = df1.append(df2)
-        expected = df1.copy()
-        assert_frame_equal(result, expected)
-
-        # Empty df with columns append empty df
-        df1 = DataFrame(columns=['bar', 'foo'])
-        df2 = DataFrame()
-        result = df1.append(df2)
-        expected = df1.copy()
-        assert_frame_equal(result, expected)
-
-        # Non-Empty df with columns append empty df
-        df1 = DataFrame(np.random.randn(5, 2), columns=['bar', 'foo'])
-        df2 = DataFrame()
-        result = df1.append(df2)
-        expected = df1.copy()
-        assert_frame_equal(result, expected)
-
-    def test_append_dtypes(self):
-
-        # GH 5754
-        # row appends of different dtypes (so need to do by-item)
-        # can sometimes infer the correct type
-
-        df1 = DataFrame({ 'bar' : Timestamp('20130101') }, index=lrange(5))
-        df2 = DataFrame()
-        result = df1.append(df2)
-        expected = df1.copy()
-        assert_frame_equal(result, expected)
-
-        df1 = DataFrame({ 'bar' : Timestamp('20130101') }, index=lrange(1))
-        df2 = DataFrame({ 'bar' : 'foo' }, index=lrange(1,2))
-        result = df1.append(df2)
-        expected = DataFrame({ 'bar' : [ Timestamp('20130101'), 'foo' ]})
-        assert_frame_equal(result, expected)
-
-        df1 = DataFrame({ 'bar' : Timestamp('20130101') }, index=lrange(1))
-        df2 = DataFrame({ 'bar' : np.nan }, index=lrange(1,2))
-        result = df1.append(df2)
-        expected = DataFrame({ 'bar' : Series([ Timestamp('20130101'), np.nan ],dtype='M8[ns]') })
-        assert_frame_equal(result, expected)
-
-        df1 = DataFrame({ 'bar' : Timestamp('20130101') }, index=lrange(1))
-        df2 = DataFrame({ 'bar' : np.nan }, index=lrange(1,2), dtype=object)
-        result = df1.append(df2)
-        expected = DataFrame({ 'bar' : Series([ Timestamp('20130101'), np.nan ],dtype='M8[ns]') })
-        assert_frame_equal(result, expected)
-
-        df1 = DataFrame({ 'bar' : np.nan }, index=lrange(1))
-        df2 = DataFrame({ 'bar' : Timestamp('20130101') }, index=lrange(1,2))
-        result = df1.append(df2)
-        expected = DataFrame({ 'bar' : Series([ np.nan, Timestamp('20130101')] ,dtype='M8[ns]') })
-        assert_frame_equal(result, expected)
-
-        df1 = DataFrame({ 'bar' : Timestamp('20130101') }, index=lrange(1))
-        df2 = DataFrame({ 'bar' : 1 }, index=lrange(1,2), dtype=object)
-        result = df1.append(df2)
-        expected = DataFrame({ 'bar' : Series([ Timestamp('20130101'), 1 ]) })
-        assert_frame_equal(result, expected)
-
     def test_asfreq(self):
         offset_monthly = self.tsframe.asfreq(datetools.bmonthEnd)
         rule_monthly = self.tsframe.asfreq('BM')
